Update type resolving for dictionaries and Any type
Expanded functionality of the type resolving methods for dictionary types and the "Any" type. For dictionaries, keys are now required to be strings, thus throwing an UnsupportedTypeError when key type is not a string. For the "Any" type, it's now resolved to an empty dictionary. Corresponding tests were added or updated accordingly.