Add functionality to ignore certain function parameters
The `function_to_schema` method now supports an optional `ignore_args` parameter that allows specific function parameters to be ignored when generating the schema. Any parameters that are listed in `ignore_args` will not be included in the generated schema. This update includes modifications to the handling of parameters as well as associated changes in tests.